Added a common function helper script
